requirements.txt
# requirements.txt — Production dependencies
faker==24.11.0
pandas==2.2.2
numpy==1.26.4
pyodbc==5.1.0
sqlalchemy==2.0.30
lightgbm==4.3.0
scikit-learn==1.5.0
mlflow==2.13.2
azure-identity==1.17.1
azure-keyvault-secrets==4.8.0
python-dotenv==1.0.1
